testing/web-platform/tests/webrtc/RTCDataChannel-close.html

<!doctype html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>RTCDataChannel.prototype.close</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
'use strict';

promise_test(async t => {
  const pc1 = new RTCPeerConnection();
  t.add_cleanup(() => pc1.close());
  const [channel1, channel2] = await createDataChannelPair(pc1);
  const haveClosed = new Promise(r => channel2.onclose = r);
  let closingSeen = false;
  channel1.onclosing = t.unreached_func();
  channel2.onclosing = () => {
    assert_equals(channel2.readyState, 'closing');
    closingSeen = true;
  };
  channel2.addEventListener('error', t.unreached_func());
  channel1.close();
  await haveClosed;
  assert_equals(channel2.readyState, 'closed');
  assert_true(closingSeen, 'Closing event was seen');
}, 'Close datachannel causes onclosing and onclose to be called');

promise_test(async t => {
  // This is the same test as above, but using addEventListener
  // rather than the "onclose" attribute.
  const pc1 = new RTCPeerConnection();
  t.add_cleanup(() => pc1.close());
  const [channel1, channel2] = await createDataChannelPair(pc1);
  const haveClosed = new Promise(r => channel2.addEventListener('close', r));
  let closingSeen = false;
  channel1.addEventListener('closing', t.unreached_func());
  channel2.addEventListener('closing', () => {
    assert_equals(channel2.readyState, 'closing');
    closingSeen = true;
  });
  channel2.addEventListener('error', t.unreached_func());
  channel1.close();
  await haveClosed;
  assert_equals(channel2.readyState, 'closed');
  assert_true(closingSeen, 'Closing event was seen');
}, 'Close datachannel causes closing and close event to be called');

promise_test(async t => {
  const pc1 = new RTCPeerConnection();
  t.add_cleanup(() => pc1.close());
  const [channel1, channel2] = await createDataChannelPair(pc1);
  const events = [];
  let error = null;
  channel2.addEventListener('error', t.step_func(event => {
    events.push('error');
    assert_true(event instanceof RTCErrorEvent);
    error = event.error;
  }));
  const haveClosed = new Promise(r => channel2.addEventListener('close', () => {
    events.push('close');
    r();
  }));
  pc1.close();
  await haveClosed;
  // Error should fire before close.
  assert_array_equals(events, ['error', 'close']);
  assert_true(error instanceof RTCError);
  assert_equals(error.name, 'OperationError');
}, 'Close peerconnection causes close event and error to be called');

</script>
